improve app start performance resolves #281
