+++ index.js
@@ -2,11 +2,7 @@ import {
   Navigation,
   ScreenVisibilityListener as RNNScreenVisibilityListener
 } from "react-native-navigation";
-import { AsyncStorage, NetInfo } from "react-native";
-import RSAKey from "react-native-rsa";
-import DeviceInfo from "react-native-device-info";
-import { sha256 } from "react-native-sha256";
-import Config from "react-native-config";
+import { NetInfo } from "react-native";
 // import Reactotron from "reactotron-react-native";
 
 import client, { persistor } from "./src/graphql/client";
@@ -16,9 +12,7 @@ import registerScreens from "./src/screens";
 
 import IS_INSTRUCTIONS_SHOWN from "./src/graphql/queries/isInstructionShown";
 import setCurrentScreen from "./src/graphql/mutations/setCurrentScreen";
-import SIGN_UP from "./src/graphql/mutations/signUp";
 import UPDATE_NETWORK_STATUS from "./src/graphql/mutations/updateNetworkStatus";
-import ME from "./src/graphql/queries/me";
 
 import topTabs from "./src/screens/VoteList/topTabs";
 
@@ -83,47 +77,7 @@ class App {
     return isInstructionsShown;
   };
 
-  getNewToken = async () => {
-    const rsa = new RSAKey();
-    rsa.setPublicString(Config.PUBLIC_KEY); // return json encoded string
-    const uniqueID = await sha256(DeviceInfo.getUniqueID());
-    const deviceHashEncrypted = rsa.encrypt(uniqueID);
-
-    try {
-      const { data } = await client.mutate({
-        mutation: SIGN_UP,
-        variables: {
-          deviceHashEncrypted
-        }
-      });
-
-      await AsyncStorage.setItem("authorization", data.signUp.token);
-    } catch (error) {
-      // TODO: Show later a message that user is not registered
-    }
-  };
-
   startApp = async ({ isInstructionsShown = false } = {}) => {
-    // await AsyncStorage.removeItem("authorization");
-    const token = await AsyncStorage.getItem("authorization");
-    if (!token) {
-      await this.getNewToken();
-    } else {
-      try {
-        const me = await client
-          .query({
-            query: ME,
-            fetchPolicy: "network-only"
-          })
-          .then(({ data }) => data.me);
-        if (!me) {
-          await this.getNewToken();
-        }
-      } catch (error) {
-        // TODO: handle this
-      }
-    }
-
     // Decide Startscreen
     if (isInstructionsShown) {
       Navigation.startSingleScreenApp({
