+++ src/graphql/client.js
@@ -8,13 +8,19 @@ import { InMemoryCache } from "apollo-cache-inmemory";
 import { CachePersistor } from "apollo-cache-persist";
 import { setContext } from "apollo-link-context";
 import { createNetworkStatusNotifier } from "react-apollo-network-status";
+import RSAKey from "react-native-rsa";
+import DeviceInfo from "react-native-device-info";
+import { sha256 } from "react-native-sha256";
+import Config from "react-native-config";
 
-import Config from "../config";
+import Configuration from "../config";
 
 import { defaults, resolvers } from "./resolvers";
 import typeDefs from "./schemas";
 
 import UPDATE_NETWORK_STATUS from "../graphql/mutations/updateNetworkStatus";
+import ME from "../graphql/queries/me";
+import SIGN_UP from "../graphql/mutations/signUp";
 
 let client; // eslint-disable-line
 
@@ -37,9 +43,50 @@ const persistor = new CachePersistor({
   debounce: 1000
 });
 
+const getNewToken = async () => {
+  const rsa = new RSAKey();
+  rsa.setPublicString(Config.PUBLIC_KEY); // return json encoded string
+  const uniqueID = await sha256(DeviceInfo.getUniqueID());
+  const deviceHashEncrypted = rsa.encrypt(uniqueID);
+
+  try {
+    const { data } = await client.mutate({
+      mutation: SIGN_UP,
+      variables: {
+        deviceHashEncrypted
+      }
+    });
+
+    await AsyncStorage.setItem("authorization", data.signUp.token);
+  } catch (error) {
+    // TODO: Show later a message that user is not registered
+  }
+};
+
+let me = null;
+
 const authLink = setContext(async (_, { headers }) => {
   // get the authentication token from local storage if it exists
   const token = await AsyncStorage.getItem("authorization");
+  if (_.operationName !== "me" && _.operationName !== "signUp") {
+    if (!token && _.operationName !== "signUp") {
+      await getNewToken();
+    } else if (!me) {
+      try {
+        me = await client
+          .query({
+            query: ME,
+            fetchPolicy: "network-only"
+          })
+          .then(({ data }) => data.me);
+        if (!me) {
+          await getNewToken();
+        }
+      } catch (error) {
+        // TODO: handle this
+      }
+    }
+  }
   // return the headers to the context so httpLink can read them
   return {
     headers: {
@@ -94,7 +141,7 @@ client = new ApolloClient({
     networkStatusNotifierLink,
     authLink,
     stateLink,
-    new HttpLink({ uri: Config.GRAPHQL_URL })
+    new HttpLink({ uri: Configuration.GRAPHQL_URL })
   ])
 });
 export default client;
